Correct default serial device settings
The new boards run their serial devices at 38400 baud by default.
While the control register is set to the correct value by WRAMPmon
during normal operation, that doesn't happen if the RexSimulator
project is used as a standalone library, such as in mark200.

The serial device was also set to use the old clockrate, which causes
the device to run as though the board was closed at 4MHz rather than
6.25MHz.