Stop crash on certain invalid escape sequences.
If the cursor position is set outside the bounds of the serial port
form, the simulator would crash with an index-out-of-bounds
error. This would cause issues with code that does not recognise
that that escape sequence is 1-indexed, such as breakout.

The serial port form now clamps the cursor position to valid bounds.